Refactor: reverted some changes
