+++ packages/protocol/engine/entrypointsmanager.go
@@ -1,4 +1,4 @@
-package eviction
+package engine
 
 import (
 	"sync"
